// types/GithubCommitResponse.ts
export interface GithubCommitResponse {
  sha: string
  node_id: string
  commit: Commit
  url: string
  html_url: string
  comments_url: string
  author: Author2
  committer: Committer2
  parents: Parent[]
  stats: Stats
  files: File[]
}

interface Commit {
  author: Author
  committer: Committer
  message: string
  tree: Tree
  url: string
  comment_count: number
  verification: Verification
}

interface Author {
  name: string
  email: string
  date: string
}

interface Committer {
  name: string
  email: string
  date: string
}

interface Tree {
  sha: string
  url: string
}

interface Verification {
  verified: boolean
  reason: string
  signature: any
  payload: any
  verified_at: any
}

interface Author2 {
  login: string
  id: number
  node_id: string
  avatar_url: string
  gravatar_id: string
  url: string
  html_url: string
  followers_url: string
  following_url: string
  gists_url: string
  starred_url: string
  subscriptions_url: string
  organizations_url: string
  repos_url: string
  events_url: string
  received_events_url: string
  type: string
  user_view_type: string
  site_admin: boolean
}

interface Committer2 {
  login: string
  id: number
  node_id: string
  avatar_url: string
  gravatar_id: string
  url: string
  html_url: string
  followers_url: string
  following_url: string
  gists_url: string
  starred_url: string
  subscriptions_url: string
  organizations_url: string
  repos_url: string
  events_url: string
  received_events_url: string
  type: string
  user_view_type: string
  site_admin: boolean
}

interface Parent {
  sha: string
  url: string
  html_url: string
}

interface Stats {
  total: number
  additions: number
  deletions: number
}

interface File {
  sha: string
  filename: string
  status: string
  additions: number
  deletions: number
  changes: number
  blob_url: string
  raw_url: string
  contents_url: string
  patch: string
}